Move declarations to lamp.h. Move rainbow effects to rainbow.h
